backend/requirements.txt
aiofiles==23.1.0
aiotinydb==2.0.0
anyio==3.7.0
certifi==2023.5.7
click==8.1.3
fastapi==0.98.0
freetype-py==2.3.0
h11==0.14.0
httpcore==0.17.2
httpx==0.24.1
idna==3.4
iniconfig==2.0.0
packaging==23.1
Pillow==9.5.0
pluggy==1.2.0
pycairo==1.24.0
pydantic==1.10.9
pyserial==3.5
pyserial-asyncio==0.6
pytest==7.4.0
pytest-asyncio==0.21.0
python-dateutil==2.8.2
reportlab==4.0.0
rlPyCairo==0.3.0
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
tinydb==4.8.0
typing_extensions==4.7.0
uvicorn==0.22.0
